+++ Environments.EnvironmentMappings.Tests/EnvironmentVariableCollection.cs
@@ -1,0 +1,8 @@
+using Xunit;
+
+namespace Environments.EnvironmentMappings.Tests;
+
+[CollectionDefinition("EnvironmentVariables", DisableParallelization = true)]
+public sealed class EnvironmentVariablesCollection
+{
+}